fix bug with never snooze

+++ Sleepy-time/Models/SleepyTime.swift
@@ -21,7 +21,7 @@ enum SleepyTimeType {
 
 
 struct MySettings {
-    var alarmVolume: Int
+    var alarmVolume: Double
     var fallAsleepTime: Int
     var isVibrated: Bool
     var ringtone: Data?
